+++ packages/orq-rc/src/models/components/evaluatorresponsejsonschema.ts
@@ -28,8 +28,8 @@ export const EvaluatorResponseJsonSchema$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-07-02T16:16:01.733Z"),
-  updated: z.string().default("2026-07-02T16:16:01.734Z"),
+  created: z.string().default("2026-07-03T05:48:17.067Z"),
+  updated: z.string().default("2026-07-03T05:48:17.067Z"),
   updated_by_id: z.nullable(z.string()).optional(),
   guardrail_config: z.any().optional(),
   type: z.literal("json_schema"),