+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -412,7 +412,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KWHSVZPCZVQPA62Z1KTW5XN0"),
+  _id: z.string().default("01KWK8B2EYVVP9J95A8SRH6GEE"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: CreateDatasourceStatus$inboundSchema,